refactor: 예외 검증 코드 모듈로 분리
- 폴더명 일관성 있게 변경
