+++ src/js/index.js
@@ -4,24 +4,3 @@ import LottoUI from './lottoUI.js';
 const lottoUI = new LottoUI();
 const lottoController = new LottoController(lottoUI);
 lottoController.init();
-
-/*
-const $showResultButton = document.querySelector('.open-result-modal-button')
-const $modalClose = document.querySelector('.modal-close')
-const $modal = document.querySelector('.modal')
-const $lottoNumbersToggleButton = document.querySelector(
-  '.lotto-numbers-toggle-button'
-)
-
-const onModalShow = () => {
-  $modal.classList.add('open')
-}
-
-const onModalClose = () => {
-  $modal.classList.remove('open')
-}
-
-$showResultButton.addEventListener('click', onModalShow)
-$modalClose.addEventListener('click', onModalClose)
-
- */
