+++ src/js/lottoController.js
@@ -1,6 +1,7 @@
-import { $ } from './util/dom.js';
-import Lotto from './object/Lotto.js';
-import { ALERT_MESSAGES, LOTTO_SETTINGS, DOM_SELECTORS } from './constants.js';
+import { $ } from './utils/dom.js';
+import Lotto from './objects/Lotto.js';
+import { ALERT_MESSAGES, LOTTO_SETTINGS, DOM_SELECTORS } from './utils/constants.js';
+import { isMoneyUnderLottoPrice, isMoneyNotInteger } from './utils/validation.js';
 
 export default class LottoController {
   constructor(lottoUI) {
@@ -26,11 +27,11 @@ export default class LottoController {
 
   handleMoneyInputButton() {
     const moneyInput = Number($(DOM_SELECTORS.MONEY_INPUT).value);
-    if (moneyInput < LOTTO_SETTINGS.LOTTO_PRICE) {
+    if (isMoneyUnderLottoPrice(moneyInput)) {
       alert(ALERT_MESSAGES.UNDER_MIN_PRICE);
       return;
     }
-    if (Math.floor(moneyInput) !== moneyInput) {
+    if (isMoneyNotInteger(moneyInput)) {
       alert(ALERT_MESSAGES.NOT_INTEGER_PRICE);
       return;
     }
