+++ src/js/objects/Lotto.js
@@ -1,4 +1,4 @@
-import { LOTTO_SETTINGS } from '../constants.js';
+import { LOTTO_SETTINGS } from '../utils/constants.js';
 
 export default class Lotto {
   constructor() {
